backend/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# AI/ML dependencies
sentence-transformers==2.2.2
pinecone-client==2.2.4
cohere==4.37

# AWS dependencies
boto3==1.34.0
botocore==1.34.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Optional: For local development
fastapi==0.104.1
uvicorn==0.24.0 